requirements.txt
GitPython
aiohttp
tqdm
gradio
python-dotenv
